requirements.txt
Flask==3.1.0
supabase==2.10.0
Pillow==11.0.0
bcrypt==4.2.1
python-dotenv==1.0.0
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
